Update the driver to make it compatible with the ESP32 SDK 2.x
The original code was calling Wire.endTransmission(false) to indicate that
the transaction has not been completed to collect the measurement data
which comes later after the measurement has been completed.
Unfortunately this does not work with the ESP32 Arduino SDK2 anymore,
the workaround is to end the transaction, wait 20ms for the measurement
and try to read the data after the delay.

Another point is that the return values for endTransmission() has been changed
7 (I2C_ERROR_CONTINUE) is not being returned anymore. The SDK returns only
0,2,5,4 and there are no defines for it.